+++ packages/dso-toolkit/src/components/date-picker/date-picker.models.ts
@@ -1,6 +1,10 @@
 export interface DatePicker {
   id?: string;
   dsoDateChange?: (e: CustomEvent) => void;
+  dsoBlur?: (e: CustomEvent) => void;
+  dsoKeyUp?: (e: CustomEvent) => void;
+  dsoKeyDown?: (e: CustomEvent) => void;
+  dsoFocus?: (e: CustomEvent) => void;
   value?: string;
   min?: string;
   max?: string;
@@ -8,4 +12,5 @@ export interface DatePicker {
   autofocus?: boolean;
   invalid?: boolean;
   describedBy?: string;
+  required?: boolean;
 }